fix: moved helper to `next/lib` to avoid having to bundle entire export dir
