+++ packages/next/src/lib/is-dynamic-usage-error.ts
@@ -1,7 +1,7 @@
-import { DynamicServerError } from '../../client/components/hooks-server-context'
-import { NotFoundError } from '../../client/components/not-found'
-import { RedirectError } from '../../client/components/redirect'
-import { DynamicNoSSRError } from '../../shared/lib/lazy-dynamic/no-ssr-error'
+import { DynamicServerError } from '../client/components/hooks-server-context'
+import { NotFoundError } from '../client/components/not-found'
+import { RedirectError } from '../client/components/redirect'
+import { DynamicNoSSRError } from '../shared/lib/lazy-dynamic/no-ssr-error'
 
 type DynamicUsageError =
   | DynamicServerError
